requirements.txt
aenum==3.1.15
ai-api-client-sdk==2.0.4
ai-core-sdk==2.1.1
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
astroid==3.0.2
async-timeout==4.0.3
attrs==23.1.0
awscli==1.32.34
billiard==4.2.0
boto3==1.34.34
botocore==1.34.34
cachetools==5.3.2
celery==5.3.6
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
dill==0.3.8
docutils==0.16
frozenlist==1.4.0
future-fstrings==1.2.0
google-auth==2.27.0
idna==3.4
isort==5.13.2
jmespath==1.0.1
kombu==5.3.4
kubernetes==29.0.0
magic-filter==1.0.12
mccabe==0.7.0
mental-api-client==1.0.2
metaflow==2.8.0
multidict==6.0.4
oauthlib==3.2.2
platformdirs==4.2.0
prompt-toolkit==3.0.41
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
pyhumps==3.8.0
pylint==3.0.3
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.10.0
sap-ai-core-metaflow==1.1.14
six==1.16.0
tomlkit==0.12.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.12
websocket-client==1.7.0
yarl==1.9.2
